bot: Undid a lot of explicit typing, see notes below database: Implemented guild preferences -> logging channel ID retrieval general: Cleaned up a lot of code that was previously forgotten about general: Changed imports to directly import needed classes and functions instead of the entire module logging: Implemented channel, guild, member, message and role logging
After some consideration, we settled on a C++-like (like as in the way it looks, not how it works) typing model, essentially only declaring the type when it would otherwise not be explicit, e.g if we're calling a constructor of a class, the type is explicitly set anyway.
Declaring the type of every variable was, due to Python's architecture, making too much code very cluttered, which is why we switched away from a typing model that declares types for all variables no matter the context.

In a previous commit, a change was made to the usage of the variable name for the command context, to be more "consistent with nextcord".
This proved to be wrong.
Among all classes, "ctx" is more widely employed as a name for the context variable, only a few classes (e.g the Help command class) use "context" instead of "ctx".
We settled on not bothering with either, but instead consistently either one across any script individually.