feat: Add raid ban command to AntiRaid cog
This commit adds a new command called `raid_ban` to the `AntiRaid` cog. The `raid_ban` command allows moderators to ban raiders in the Discord server. It takes up to 10 raider users as arguments and performs the ban action. The command also interacts with the database to store information about the ban cases. This feature enhances the functionality of the AntiRaid cog and helps in effectively dealing with raiders.

#210